Adding composition based structure
Adding urls as environment vars and settings
Controlling batch sizes from python code
