+++ elasticassignment/ageofempires/management/commands/index_units.py
@@ -1,8 +1,10 @@
 from django.core.management.base import BaseCommand
-import ageofempires.management.commands.command_classes as command_classes
+from django.conf import settings
+import ageofempires.models.units as Units
 
 class Command(BaseCommand):
 
     def handle(self, *args, **kwargs):
-        command_classes.Units.make_index()
-        command_classes.Units.populate_index("https://age-of-empires-2-api.herokuapp.com/api/v1/units")
+        url = settings.AOE_BASE_URL + settings.DATA_URLS['UNITS']
+        Units.Units.make_index()
+        Units.Units.populate_index(url)
