+++ elasticassignment/elasticassignment/.env.example
@@ -1,0 +1,4 @@
+ELASTIC_HOST=127.0.0.1
+NUMBER_OF_SHARDS=1
+NUMBER_OF_REPLICAS=0
+AOE_BASE_URL=https://age-of-empires-2-api.herokuapp.com
